server/codeshipper/requirements.txt
django==2.2
requests
PyJWT==1.6.4
pytz
